+++ common/src/main/java/net/mehvahdjukaar/polytone/colormap/IColormapNumberProvider.java
@@ -37,10 +37,6 @@ public interface IColormapNumberProvider {
         return true;
     }
 
-    default boolean usesRandom(){
-        return false;
-    }
-
     default IColormapNumberProvider createConcurrent() {
         return this;
     }
@@ -153,11 +149,6 @@ public interface IColormapNumberProvider {
         @Override
         public boolean usesState() {
             return false;
-        }
-
-        @Override
-        public boolean usesRandom() {
-            return true; //uses random
         }
     });