+++ changelog.md
@@ -1,2 +1,3 @@
 - optimized expressions to just use a concurrent approach when used by block colors. This means that all non blocks colormaps that use expressions should be much faster now
 - fixed cloud height expression
+- fixed fluid modifiers not using concurrent colormaps